refactor: InputContainer 컴포넌트 분리 및 변수명 수정
