+++ src/components/lotto/WinningLottoPanel/styles.ts
@@ -14,21 +14,6 @@ export const GuideMessage = styled.p`
   margin-bottom: 20px;
 `;
 
-export const NumberContainer = styled.div`
-  display: flex;
-  flex-direction: column;
-`;
-
-interface NumberContainerStyleProps {
-  $inputStyle?: React.CSSProperties;
-}
-
-export const NumberInputContainer = styled.div<NumberContainerStyleProps>`
-  display: flex;
-  gap: 5px;
-
-  ${({ $inputStyle }) => $inputStyle && { ...$inputStyle }};
-`;
 
 export const NumberLabel = styled.label`
   font-size: 12px;
