+++ src/components/lotto/index.tsx
@@ -5,3 +5,4 @@ export { default as MoneyPanel } from './MoneyPanel';
 export { default as WinningLottoPanel } from './WinningLottoPanel';
 export { default as ResultModal } from './ResultModal';
 export { default as Lotto } from './Lotto';
+export { default as InputContainer } from './InputContainer';
